feat(rework-todolist-api-call): put AI functionality in his own component
