+++ client/src/views/ToDoList/ToDoListDetailView.vue
@@ -2,17 +2,22 @@
 import { ref, computed, onMounted, onUnmounted } from 'vue';
 import { useRoute } from 'vue-router';
 
-import { client } from '@/services/requestMaker.js';
-import { hookApi } from '@/services/requestHook.js';
-import { setTitle, setDescription } from "@/utils/documentInfos.js";
 import ToDoItemFormComponent from '@/components/ToDoList/ToDoItemFormComponent.vue';
 import ToDoItemImageModalComponent from '@/components/ToDoList/ToDoItemImageModalComponent.vue';
 import ToggleComponent from '@/components/ToggleComponent.vue';
 import LoaderComponent from '@/components/Loader/LoaderComponent.vue';
 import QRCodeModal from '@/components/ToDoList/ToDoListQRCodeModal.vue';
+import AIGenerateTasks from '@/components/ToDoList/AIGenerateTasks.vue';
+import { setTitle, setDescription } from "@/utils/documentInfos.js";
 import logger from '@/utils/logger.js';
+import { client } from '@/services/requestMaker.js';
+import { hookApi } from '@/services/requestHook.js';
+import { ToDoItemService } from "@/services/toDoItemService.js";
+import { ToDoListService } from "@/services/toDoListService.js";
 
 const route = useRoute();
+const toDoItemService = new ToDoItemService();
+const toDoListService = new ToDoListService();
 const { isLoading, error, executeRequest } = hookApi();
 
 const toDoList = ref({});
@@ -28,11 +33,6 @@ const qrCodeUrl = ref(null);
 const linkUrl = ref(null);
 const showItemImageModal = ref(false);
 const openMenuId = ref(null);
-
-// Pour l'IA
-const aiPrompt = ref('');
-const isGeneratingWithAI = ref(false);
-const aiError = ref(null);
 
 // Récupération des items de la ToDoList
 const fetchToDoItems = async () => {
@@ -192,10 +192,6 @@ const toggleMenu = (item) => {
   openMenuId.value = openMenuId.value === item.id ? null : item.id;
 };
 
-// const closeMenu = (item) => {
-//   item.showMenu = false;
-// };
-
 // Fonction pour fermer tous les menus
 const closeAllMenus = (event) => {
   // Vérifie si le clic est à l'intérieur du menu ou du bouton (évite la fermeture immédiate)
@@ -216,32 +212,8 @@ onUnmounted(() => {
   document.removeEventListener('click', closeAllMenus);
 });
 
-// Fonction pour générer des tâches avec l'IA
-const generateWithAI = async () => {
-  if (!aiPrompt.value.trim() || isGeneratingWithAI.value) return;
-
-  isGeneratingWithAI.value = true;
-  aiError.value = null;
-  try {
-    const generatedTasks = await executeRequest(() => client.post('/api/ai/generate-todolist', { prompt: aiPrompt.value }));
-    if (generatedTasks && generatedTasks.length > 0) {
-      // Ajouter chaque tâche générée comme un nouvel item
-      for (const taskTitle of generatedTasks) {
-        const newItem = { title: taskTitle, done: false };
-        // Appel direct pour créer l'item (similaire à ce qui est fait dans ToDoItemFormComponent)
-        const response = await executeRequest(
-            () => client.postWithFile(`/api/todolist/${route.params.id}/todoitem`, newItem)
-        );
-        toDoItems.value.push(response.toDoItem);
-      }
-      aiPrompt.value = ''; // Vider le champ après la génération
-    }
-  } catch (err) {
-    logger.error('Error generating tasks with AI:', err?.response?.data?.message || err.message);
-    aiError.value = err?.response?.data?.message || 'Erreur lors de la génération des tâches.';
-  } finally {
-    isGeneratingWithAI.value = false;
-  }
+const handleTasksGenerated = (newItems) => {
+  toDoItems.value.push(...newItems);
 };
 
 </script>
@@ -279,34 +251,7 @@ const generateWithAI = async () => {
       </div>
 
       <!-- Section IA -->
-      <div class="px-4 mb-6">
-        <div class="bg-white dark:bg-gray-800 p-4 rounded-xl shadow-lg dark:shadow-gray-700">
-          <h3 class="text-lg font-semibold mb-2 text-gray-700 dark:text-gray-300 flex items-center">
-            <v-icon name="ri-robot-line" class="mr-2" scale="1.2" />
-            Générer des tâches avec l'IA
-          </h3>
-          <div class="flex items-center gap-2">
-            <input
-                v-model="aiPrompt"
-                type="text"
-                placeholder="Ex: Ingrédients pour une tarte aux pommes"
-                class="flex-grow border border-gray-300 dark:border-gray-600 p-3 rounded-lg focus:outline-none focus:ring-2 focus:ring-blue-600 dark:focus:ring-yellow-400 bg-white dark:bg-gray-700 text-gray-900 dark:text-white"
-                @keyup.enter="generateWithAI"
-            />
-            <button
-                class="bg-green-500 hover:bg-green-600 text-white px-4 py-3 rounded-lg flex items-center justify-center transition duration-300"
-                :disabled="isGeneratingWithAI || !aiPrompt.trim()"
-                @click="generateWithAI"
-            >
-              <v-icon v-if="!isGeneratingWithAI" name="io-sparkles-outline" scale="1.2"/>
-              <LoaderComponent v-else :small="true" />
-              <span class="ml-2 hidden sm:inline">{{ isGeneratingWithAI ? 'Génération...' : 'Générer' }}</span>
-            </button>
-          </div>
-          <p v-if="aiError" class="text-sm mt-2 text-red-600 dark:text-red-400">{{ aiError }}</p>
-        </div>
-      </div>
-
+      <AIGenerateTasks v-if="toDoList.id" :to-do-list-id="toDoList.id" @tasks-generated="handleTasksGenerated" />
 
       <!-- QRCodeModal -->
       <QRCodeModal
